requirements.txt

torch
torchvision
torchaudio
diffusers
transformers
streamlit
Pillow
